// lambda-ts/types/generated/models/SessionResponseSessionsInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 「make it !」REST API仕様書
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SessionResponseSessionsInner
 */
export interface SessionResponseSessionsInner {
    /**
     * セッション ID
     * @type {string}
     * @memberof SessionResponseSessionsInner
     */
    sessionId: string;
    /**
     * セッションタイトル
     * @type {string}
     * @memberof SessionResponseSessionsInner
     */
    sessionTitle: string;
    /**
     * 転送されたセッションかどうか
     * @type {boolean}
     * @memberof SessionResponseSessionsInner
     */
    isEscalated: boolean;
    /**
     * メッセージ開始日時（ISO8601 形式）
     * @type {string}
     * @memberof SessionResponseSessionsInner
     */
    createdAt: string;
    /**
     * メッセージ更新日時（ISO8601 形式）
     * @type {string}
     * @memberof SessionResponseSessionsInner
     */
    updatedAt: string;
}

/**
 * Check if a given object implements the SessionResponseSessionsInner interface.
 */
export function instanceOfSessionResponseSessionsInner(value: object): value is SessionResponseSessionsInner {
    if (!('sessionId' in value) || value['sessionId'] === undefined) return false;
    if (!('sessionTitle' in value) || value['sessionTitle'] === undefined) return false;
    if (!('isEscalated' in value) || value['isEscalated'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function SessionResponseSessionsInnerFromJSON(json: any): SessionResponseSessionsInner {
    return SessionResponseSessionsInnerFromJSONTyped(json, false);
}

export function SessionResponseSessionsInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionResponseSessionsInner {
    if (json == null) {
        return json;
    }
    return {
        
        'sessionId': json['sessionId'],
        'sessionTitle': json['sessionTitle'],
        'isEscalated': json['isEscalated'],
        'createdAt': json['createdAt'],
        'updatedAt': json['updatedAt'],
    };
}

export function SessionResponseSessionsInnerToJSON(value?: SessionResponseSessionsInner | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'sessionId': value['sessionId'],
        'sessionTitle': value['sessionTitle'],
        'isEscalated': value['isEscalated'],
        'createdAt': value['createdAt'],
        'updatedAt': value['updatedAt'],
    };
}